+++ src/natten/backends/configs/checks.py
@@ -243,7 +243,7 @@ def can_run_cutlass_hopper_fmha(
         key,
         value,
         must_match_head_dims=True,
-        supports_gqa_mqa=False,
+        supports_gqa_mqa=True,  # NOTE: supports GQA in only by virtue of repeating heads manually
         raise_error=raise_error,
         backend_name="Hopper FMHA",
     ):
@@ -335,7 +335,7 @@ def can_run_cutlass_hopper_fna(
         key,
         value,
         must_match_head_dims=True,
-        supports_gqa_mqa=False,
+        supports_gqa_mqa=True,  # NOTE: supports GQA in only by virtue of repeating heads manually
         raise_error=raise_error,
         backend_name="Hopper FNA",
     ):